libdvbv5: move ISDB-T frq conversion to the descriptors code
Instead of having this done on several places, just do it once,
at the code that parses the ISDB-T descriptor.

This is the same as the other descriptors do.

+++ lib/include/libdvbv5/desc_isdbt_delivery.h
@@ -27,7 +27,7 @@
 struct isdbt_desc_terrestrial_delivery_system {
 	DVB_DESC_HEADER();
 
-	uint16_t *frequency;
+	uint32_t *frequency;
 	unsigned num_freqs;
 
 	union {
